moved PFBend components into own folder
